Rewrote most of the Wordpress extension.  The following are changes of note.
  - Upgraded to Wordpress 4.0.
  - Added support for sshfs. More on this below.
  - Write Wordpress version to app/.wordpress.version so that users can
    easily find and monitor it.
  - Now uses WEBDIR instead of hard coding `htdocs`.

Using `sshfs` has the following requirements.

  - You must include a `.ssh` directory at the root of your project.
  - This folder must include a public and private key.
  - You must add the public key to the `authorized_keys` file on your
    SSH server.
  - You must also add a `known_hosts` file to the `.ssh` folder.  This
    needs to have a record for your remote server.
  - You need to set `SSH_HOST`, `SSH_KEY_NAME` AND `SSH_PATH` in either
    your `manifest.yml` file or in `.bp-config/options.json`.
  - `SSH_HOST` points to your ssh server (ip or dns)
  - `SSH_PATH` points to the path on your ssh server to mount
  - `SSH_KEY_NAME` is the name of the key you put in the `.ssh`
    directory.

The support for sshfs works like this.  When `SSH_HOST`, `SSH_KEY_NAME`
AND `SSH_PATH` are detected by the build pack, it will mount the path
from the remote server to `app/htdocs/wp-content` in the application
container using the `sshfs` command.

If necessary, the build pack will initialize the remote directory with
minimal files required for `wp-content`.  If files exist in this
location already, it will not modify or remove them.  This is because
Wordpress will write files to this directory (themes, plugins and
uploads) and those files are left in place.  As mentioned, the build
pack does this *if necessary*, which is on the first push, on Wordpress
upgrade or if `WORDPRESS_LOCKED` is set to `False` (defaults to
`True`).

The build pack monitors the remote file system (i.e. the sshfs command)
and if the remote file system is disconnected, it should trigger the
application to restart and reconnect to the remote file system.

Finally, you will be able to see the files from your application running
on CloudFoundry by looking at the file system on the remote server.  *DO
NOT EDIT THESE FILES OR THE DIRECTORY*.  Your changes will be
overwritten and probably at a very inconvenient time.